Update config handling to Radicale 3.0
Radicale no longer exposes the configparser directly, but combines
configuration from multiple sources and merges this in its own
Configuration object, with a different API (no has_option, no
getboolean).

Note that bool conversion now happens on config load based on config
schema and there seems to be no way for plugins to add items to the
schema. To ensure consistent boolean parsing, this calls a radicale
internal function to convert the unparsed string value to a bool.